move CalendarEvent to its own file
